+++ Sources/Storage/UBUserDefault/UBCodable.swift
@@ -1,0 +1,38 @@
+//
+//  UBCodable.swift
+//  UBFoundation
+//
+//  Created by Zeno Koller on 31.01.20.
+//
+
+import Foundation
+
+public typealias UBCodable = UBDecodable & UBEncodable
+
+public protocol UBEncodable: Encodable, UBUserDefaultsStorable {}
+
+extension UBEncodable {
+    public func store(in userDefaults: UserDefaults, key: String) {
+        let data = try? JSONEncoder().encode(self)
+        userDefaults.set(data, forKey: key)
+    }
+}
+
+public protocol UBDecodable: Decodable, UBUserDefaultsRetrievable {}
+
+extension UBDecodable {
+    public static func retrieve(from userDefaults: UserDefaults, key: String, defaultValue: Self) -> Self {
+        guard let data = userDefaults.object(forKey: key) as? Data else {
+            return defaultValue
+        }
+        let value = try? JSONDecoder().decode(Self.self, from: data)
+        return value ?? defaultValue
+    }
+
+    public static func retrieveOptional(from userDefaults: UserDefaults, key: String) -> Self? {
+        guard let data = userDefaults.object(forKey: key) as? Data else {
+            return nil
+        }
+        return try? JSONDecoder().decode(Self.self, from: data)
+    }
+}