+++ Sources/UI Elements/UIView/UIView+Shadows.swift
@@ -1,0 +1,23 @@
+//
+//  UIView+Shadows.swift
+//  UBFoundation iOS
+//
+//  Created by Marco Zimmermann on 23.09.19.
+//
+
+import UIKit
+
+// MARK: - UIView Shadow Extensions
+
+extension UIView {
+    /// Adds shadow to UIView with black color and other parameters
+    public func ub_addShadow(with color: UIColor = UIColor.black, radius: CGFloat, opacity: CGFloat, xOffset: CGFloat, yOffset: CGFloat) {
+        layer.shadowColor = color.cgColor
+        layer.shadowOpacity = Float(opacity)
+        layer.shadowOffset = CGSize(width: xOffset, height: yOffset)
+        layer.shadowRadius = radius
+        layer.masksToBounds = false
+        layer.shouldRasterize = true
+        layer.rasterizationScale = UIScreen.main.scale
+    }
+}